requirements.txt
fastapi==0.111.0
redis==5.0.3
uvicorn[standard]
python-dotenv
asyncpg
# requirements.txt is for Python dependencies only — it’s used by pip to install packages inside your FastAPI app container.